change biginteger to long for better performance

+++ src/main/java/com/example/employee/model/dto/DepartmentResponse.java
@@ -3,12 +3,10 @@ package com.example.employee.model.dto;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 @Builder
 public class DepartmentResponse {
-    private BigInteger id;
+    private long id;
     private String name;
     private String description;
