+++ src/main/java/com/example/employee/model/dto/EmployeeResponse.java
@@ -3,13 +3,12 @@ package com.example.employee.model.dto;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
 @Builder
 public class EmployeeResponse {
-    private BigInteger id;
+    private long id;
     private String name;
     private String jobTitle;
