+++ src/main/java/com/example/employee/repository/DepartmentRepository.java
@@ -4,8 +4,6 @@ import com.example.employee.model.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface DepartmentRepository extends JpaRepository<Department, BigInteger> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
