+++ src/main/java/com/example/employee/repository/EmployeeRepository.java
@@ -4,8 +4,6 @@ import com.example.employee.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, BigInteger> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
